Import only supported locales for date localization
