+++ components/src/DatePicker/DatePickerHeader.js
@@ -1,13 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { format } from "date-fns";
-import * as allLocales from "date-fns/locale";
 
 import theme from "../theme";
 
 import { Flex } from "../Flex";
 import { Text } from "../Type";
 import { ControlIcon } from "../Button";
+import { supportedDateLocales } from "../utils/datePickerLocales";
 
 const DatePickerHeader = ({
   date,
@@ -25,7 +25,7 @@ const DatePickerHeader = ({
       disabled={prevMonthButtonDisabled}
     />
     <Text fontWeight="bold" color={theme.colors.blackBlue} px="x1" ml="half" fontSize="large">
-      {format(date, "MMMM yyyy", { locale: allLocales[locale] })}
+      {format(date, "MMMM yyyy", { locale: supportedDateLocales[locale] })}
     </Text>
     <ControlIcon
       icon="rightArrow"
