+++ components/src/MonthRange/MonthRange.story.js
@@ -2,9 +2,9 @@ import React from "react";
 import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import { withKnobs, select } from "@storybook/addon-knobs";
-import * as allLocales from "date-fns/locale";
 
 import MonthRange from "./MonthRange";
+import { supportedDateLocales } from "../utils/datePickerLocales";
 
 storiesOf("MonthRange", module)
   .addDecorator(withKnobs)
@@ -70,6 +70,6 @@ storiesOf("MonthRange", module)
       onRangeChange={action("range changed")}
       onStartDateChange={action("start date changed")}
       onEndDateChange={action("end date changed")}
-      locale={select("locale", Object.keys(allLocales), "uk", "locales")}
+      locale={select("locale", Object.keys(supportedDateLocales), "pl", "locales")}
     />
   ));
