--- components/src/utils/RegisterDatePickerLocales.js
@@ -1,5 +1,0 @@
-import { registerLocale } from "react-datepicker";
-import * as allLocales from "date-fns/locale";
-
-export const registerDatePickerLocales = () =>
-  Object.keys(allLocales).map(locale => registerLocale(locale, allLocales[locale]));
