+++ components/src/utils/datePickerLocales.js
@@ -1,0 +1,7 @@
+import { registerLocale } from "react-datepicker";
+import { enUS, de, fr, es, ptBR, da, pl } from "date-fns/locale";
+
+export const supportedDateLocales = { enUS, de, fr, es, ptBR, da, pl };
+
+export const registerDatePickerLocales = () =>
+  Object.keys(supportedDateLocales).map(locale => registerLocale(locale, supportedDateLocales[locale]));
