+++ docs/src/pages/components/month-picker.js
@@ -17,6 +17,7 @@ import {
   PropsTable
 } from "../../components";
 import { inputFieldProps } from "../../shared/inputProps";
+import localeProp from "../../shared/localeProp";
 
 const propsRows = [
   {
@@ -56,13 +57,7 @@ const propsRows = [
     defaultValue: "undefined",
     description: "The latest date that can be selected"
   },
-  {
-    name: "locale",
-    type: "string",
-    defaultValue: "undefined",
-    description:
-      "Locale string ex: 'uk'. For all supported locales see the knobs under the 'custom locale' example in storybook"
-  }
+  localeProp
 ];
 
 export default () => (
