+++ docs/src/pages/components/time-picker.js
@@ -17,6 +17,7 @@ import {
   PropsTable
 } from "../../components";
 import selectProps from "../../shared/selectProps";
+import localeProp from "../../shared/localeProp";
 
 const propsRows = [
   {
@@ -44,13 +45,7 @@ const propsRows = [
     defaultValue: "undefined",
     description: "The latest time that can be selected."
   },
-  {
-    name: "locale",
-    type: "string",
-    defaultValue: "undefined",
-    description:
-      "Locale string ex: 'uk'. For all supported locales see the knobs under the 'custom locale' example in storybook"
-  },
+  localeProp,
   ...selectProps.filter(prop => prop.name !== "options")
 ];
