+++ docs/src/shared/localeProp.js
@@ -1,0 +1,9 @@
+const localeProp = {
+  name: "locale",
+  type: "string",
+  defaultValue: "undefined",
+  description:
+    "One of the supported locale strings: 'enUS' (US English), 'de' (German), 'fr' (French), 'es' (Spanish), 'ptBR' (Brazil Portuguese), 'da' (Dutch), 'pl' (Polish)"
+};
+
+export default localeProp;
